% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cache_info}
\alias{cache_info}
\title{Display information about the openesm cache}
\usage{
cache_info()
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Shows the location and total size of the local file cache.
}
\examples{
\donttest{
# view cache information
cache_info()
}
}
