% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{cite.openesm_dataset}
\alias{cite.openesm_dataset}
\title{Get citation information for dataset}
\usage{
\method{cite}{openesm_dataset}(x, format = "bibtex", ...)
}
\arguments{
\item{x}{An object of class \code{openesm_dataset}.}

\item{format}{Character string specifying citation format. Currently only "bibtex" is supported.}

\item{...}{Additional arguments (currently unused).}
}
\value{
Character string with citation information, returned invisibly.
}
\description{
Get citation information for dataset
}
\examples{
\donttest{
dataset <- get_dataset("0001")
cite(dataset)
}
}
