% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clear_cache}
\alias{clear_cache}
\title{Clear the openesm cache}
\usage{
clear_cache(force = FALSE)
}
\arguments{
\item{force}{Logical. If \code{TRUE}, will not ask for confirmation before
deleting. Default is \code{FALSE}.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Removes all cached openesm data from your local machine.
}
\examples{
\donttest{
# clear cache with confirmation prompt
if(interactive()) {
 clear_cache()
}

# force clear without confirmation
clear_cache(force = TRUE)
}
}
