% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{notes.openesm_dataset}
\alias{notes.openesm_dataset}
\title{Get additional notes for dataset}
\usage{
\method{notes}{openesm_dataset}(x, ...)
}
\arguments{
\item{x}{An object of class \code{openesm_dataset}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
Character vector with dataset notes, returned invisibly.
}
\description{
Get additional notes for dataset
}
\examples{
\donttest{
dataset <- get_dataset("0001")
notes(dataset)
}
}
