% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{notes}
\alias{notes}
\title{Additional notes for openESM datasets}
\usage{
notes(x, ...)
}
\arguments{
\item{x}{An object of class \code{openesm_dataset}.}

\item{...}{Additional arguments passed to methods}
}
\value{
Additional notes and information about the dataset(s)
}
\description{
Additional notes for openESM datasets
}
\examples{
\donttest{
# Get notes for a dataset
dataset <- get_dataset("0001")
notes(dataset)
}
}
