% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openesm-package.R
\docType{package}
\name{openesm-package}
\alias{openesm}
\alias{openesm-package}
\title{openesm: Access the Open Experience Sampling Method Database}
\description{
Provides programmatic access to the Open Experience Sampling Method ('openESM') database (\url{https://openesmdata.org}), a collection of harmonized experience sampling datasets. The package enables researchers to discover, download, and work with the datasets while ensuring proper citation and license compliance.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/openesm-project/openesm-r}
  \item Report bugs at \url{https://github.com/openesm-project/openesm-r/issues}
}

}
\author{
\strong{Maintainer}: Björn S. Siepe \email{bjoernsiepe@gmail.com} (\href{https://orcid.org/0000-0002-9558-4648}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Matthias Kloft \email{kloft.dev@gmail.com} (\href{https://orcid.org/0000-0003-1845-6957}{ORCID})
}

}
\keyword{internal}
