% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{print.openesm_dataset_list}
\alias{print.openesm_dataset_list}
\title{Print method for a list of openesm_dataset objects}
\usage{
\method{print}{openesm_dataset_list}(x, ...)
}
\arguments{
\item{x}{An object of class \code{openesm_dataset_list}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
Invisibly returns the original object.
}
\description{
Print method for a list of openesm_dataset objects
}
\examples{
\donttest{
datasets <- get_dataset(c("0001", "0002"))
print(datasets)
}
}
